+++ demo/src/main/java/com/pasteleria_mil_sabores/demo/model/Rol.java
@@ -9,10 +9,12 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Rol {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
+    // Ejemplo de nombres: ROLE_USER, ROLE_ADMIN
     @Column(nullable = false, unique = true, length = 50)
     private String nombre;
 }
