+++ demo/src/main/java/com/pasteleria_mil_sabores/demo/repository/CarritoRepository.java
@@ -5,6 +5,6 @@ import com.pasteleria_mil_sabores.demo.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface CarritoRepository extends JpaRepository<Carrito, Integer> {
+public interface CarritoRepository extends JpaRepository<Carrito, Long> {
     Optional<Carrito> findByUsuario(Usuario usuario);
 }
