+++ demo/src/main/java/com/pasteleria_mil_sabores/demo/repository/ItemCarritoRepository.java
@@ -3,4 +3,4 @@ package com.pasteleria_mil_sabores.demo.repository;
 import com.pasteleria_mil_sabores.demo.model.ItemCarro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemCarritoRepository extends JpaRepository<ItemCarro, Integer> { }
+public interface ItemCarritoRepository extends JpaRepository<ItemCarro, Long> { }
