+++ demo/src/main/java/com/pasteleria_mil_sabores/demo/repository/ProductoRepository.java
@@ -3,4 +3,4 @@ package com.pasteleria_mil_sabores.demo.repository;
 import com.pasteleria_mil_sabores.demo.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> { }
+public interface ProductoRepository extends JpaRepository<Producto, Long> { }
